Fix test file name
